CMakeLists.txt: use the shared library suffix with modules
This is specifically for MacOS, where cmake uses the suffix .so for
modules to be loaded in run-time, and .dylib for shared libraries to
be linked with, while OpenSSL expects .dylib for both.

With this in place, we don't need the 'ln -s' hack in our Github
actions any more.